feat: cap framerate while window is unfocused
