+++ Assets/Script/Persistent/FocusHandler.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 19ef870a41104f6ba594d5912281e605
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
