+++ Assets/Script/Persistent/GlobalVariables.cs
@@ -107,54 +107,6 @@ namespace YARG
             LoadScene(SceneIndex.Menu);
         }
 
-        // Tracks whether audio was muted because the window lost focus,
-        // so it can be restored when focus returns.
-        private bool _mutedFromFocusLoss;
-
-        // Framerate the engine is limited to while the window is unfocused.
-        private const int BackgroundFrameRate = 10;
-
-        // Tracks whether the framerate was limited because the window lost focus,
-        // so it can be restored when focus returns.
-        private bool _framerateLimitedFromFocusLoss;
-
-        private void OnApplicationFocus(bool hasFocus)
-        {
-            if (!hasFocus)
-            {
-                if (SettingsManager.Settings.MuteOnFocusLoss.Value && !_mutedFromFocusLoss)
-                {
-                    GlobalAudioHandler.SetMasterVolume(0);
-                    _mutedFromFocusLoss = true;
-                }
-
-                if (SettingsManager.Settings.LimitBackgroundFramerate.Value && !_framerateLimitedFromFocusLoss)
-                {
-                    // VSync would otherwise pin the framerate to the monitor's refresh
-                    // rate, ignoring targetFrameRate, so disable it while unfocused.
-                    QualitySettings.vSyncCount = 0;
-                    Application.targetFrameRate = BackgroundFrameRate;
-                    _framerateLimitedFromFocusLoss = true;
-                }
-            }
-            else
-            {
-                if (_mutedFromFocusLoss)
-                {
-                    GlobalAudioHandler.SetMasterVolume(SettingsManager.Settings.MasterMusicVolume.Value);
-                    _mutedFromFocusLoss = false;
-                }
-
-                if (_framerateLimitedFromFocusLoss)
-                {
-                    // Restore from the live settings so changes made while unfocused are respected.
-                    QualitySettings.vSyncCount = SettingsManager.Settings.VSync.Value ? 1 : 0;
-                    Application.targetFrameRate = SettingsManager.Settings.FpsCap.Value;
-                    _framerateLimitedFromFocusLoss = false;
-                }
-            }
-        }
-
 #if UNITY_EDITOR
 
         // For respecting the editor's mute button
